Made considerable overhaul to Location and others: Location doesn't have a 'space type' anymore. Instead, there's now a 'Politics' attribute instead. I also made it so only non-empty attributes show in LocationShowcase and CharacterShowcase.
ImagePanel now resizes itself to fit image within, and mantains the same height instead of width,
thus making it look good with both vertical and horizontal images.

LocationCreator and CharacterCreator are both larger now, so it's better to write
attributes in them. LocationShowcase's constructor also accepts more parameters
now, like a wxPoint and wxSize (which makes it possible to use the FromDIP() function
for their sizes.) They both also create the colors used in advance, so that the same
colors aren't being constructed all the time. Same with fonts.

OutlineFiles was also slightly changed to make the images look better and
different attributes are separated with more space in the buffer (\n\n\n\n).
There's also a simple 'separator' string in the generateCharacterBuffer() which isn't
really worth explaining.