Started managing memory myself more efficiently by creating the following funtions in Chapter: LoadSceneBuffers() and ClearSceneBuffers(). This means that in amProjectManager::LoadScenes() you can now opt not to load in the wxRichTextBuffers and that's the default.
I also removed the "Sex" column from amElementsNotebook::m_charList. It didn't seem to fit.

For my own sanity I changed amStoryWriter::m_chapterPos to m_chapterIndex to make it clear
to me that that number represents the wxVector index and not the chapter position in the
book (the position starts at 1, the index starts at 0). I also store a m_sectionIndex and
that already sets up the functionality of loading chapters and basically everything in
amStoryWriter. There were lots of little layout changes to the overall look and feel of it
as well, which aren't really important right now. I made sure that all documents are saved
correctly and as far as I can tell it's working perfectly. The code is also less cluttered
and more cohesive now. There are less double function calls and stuff like that. Everything
is just cleaner. m_storyView is working perfectly as well.