Another commit with loads of things hehe:     - Changed amTreeModelNodePtrArray to a standard wxVector<amTreeModelNode*>, having gotten rid of all pre-processor-made lists with wxWidgets (they state they'll be deprecated in the future). There were little additions to amDataViewModel as well, but nothing major, stuff that I didn't realize could be in the base class.     - Got rid of some random 'auto' keywords and added "wxmemdgb.h" to ALL cpp files. I think.     - Made some more changes yet again to AutoWrapTextShape::DoCalcWrappedText()'s calculations. I think I'll try to optimize it until forever, I just can't help it...     - Got rid of amElementsNotebook's destructor and just assigned the image lists to their respective wxListViews instead of deleting them myself. Because why not?
////////////////////////////////////////////////////////////////////////////////////////////////////////////

    - Okay, now comes the most important part of the commit: The Timline.
    - I changed the naming convention to amTL*******, just to follow suit with wxSF******.
    - Changed wxVector<amTLThread> and wxVector<amTLSection> to wxVector<amTLThread*> and wxVector<amTLSection*>,
giving me actual security to pass pointers around since the wxVector doesn't reallocate
the acutal threads and sections, just the pointers to them. I'm probably gonna do this to
chapters as well (instead of having a RedeclareChaptersInElements() function). Obviously both threads
and sections are now allocated on the heap.
    - Renamed everything that needed to be renamed to '***Index', for better understanding of the code
(making it explicit that the count starts from 0).
    - Started actually implementing the Timeline sidebar, having made thumbnails for Threads,
Cards and Sections. There's an amTLThumbnail, which derives from wxSFThumbnail, which substitutes the
OnPaint functionality and both the Thread and Section thumbnails derive from it. They, on the other hand,
override the DrawContent(wxDC& dc) function. The Card thumbnail derives directly from wxSFThumbnail
because it needs its own OnPaint function.
    - The buttons to add new Threads and new Sections don't work yet, but the buttons to add Cards
to Sections and Threads do work.
    - To set the data to each thumbnail (and, shortly, more stuff), is done through a lot of function calls.
the amTLTimelineCanvas checks if anything was selected in the OnLeftDown() function, then it sends the
data to amTLTimeline, which in turn sends it to the sidebar and that sets it to each thumbnail and in
the future more stuff.
    - There's an UpdateSelectedThreadSBData() and UpdateSelectedSectionSBData(), which decide which
Cards should be drawn on the thumbnails.
    - There's an amTLTimelineCanvas::GetGoodSize(), which returns the size needed for displaying the entire
Timeline, with all Sections, Threads and Cards, ignoring scale. This is used in amTLTimelineCanvas::OnUpdateVirtualSize()
and in the thumbnails.
    - Okay, I guess that's about it, but I wrote a lot more code than it sounds! Hehe