Got completely rid of the amProjectManager class and switched to a procedural coding style instead of OOP. It works 100x better and it's so much more convenient. I also changed most things to be part of the 'am' namespace instead of prefixing the class names with 'am'. GUI classes (derived from wxWindow, for example), are still prefixed with 'am' instead of being in the namespace, just to follow the wxWidgets convention. Got rid of a lot of bugs.
There's now an amBookWizard, which is used for creating new am::Books. It's already working. Next up is making
deleting and editing books possible.

This commit is huge but I'm gonna keep the message like this, so it's not overwhelming (I don't even know what
I'd write to cover everything). Most is covered in "fixed a lot of bugs" hahahaha.